+++ source/state/class_state_finite.h
@@ -40,8 +40,8 @@ class class_state_finite {
     using VType = Eigen::Matrix<Scalar, Eigen::Dynamic, 1>;
     template<auto rank>
     using TType                    = Eigen::Tensor<Scalar, rank>;
-    int                 num_sweeps = 0;
-    int                 num_moves  = 0;
+    size_t              num_sweeps = 0;
+    size_t              num_moves  = 0;
     int                 direction  = 1;
     std::optional<long> chi_lim;
     std::optional<long> chi_max;
@@ -65,15 +65,15 @@ class class_state_finite {
     const TType<1> &current_bond() const;
     void            do_all_measurements();
 
-    int  get_sweeps() const;
-    int  reset_sweeps();
-    void set_sweeps(int num_sweeps_);
-    void increment_sweeps();
+    size_t  get_sweeps() const;
+    size_t  reset_sweeps();
+    void    set_sweeps(int num_sweeps_);
+    void    increment_sweeps();
 
-    int  get_moves() const;
-    int  reset_moves();
-    void set_moves(int num_moves_);
-    void increment_moves();
+    size_t  get_moves() const;
+    size_t  reset_moves();
+    void    set_moves(int num_moves_);
+    void    increment_moves();
 
     long                   get_chi_lim() const;
     void                   set_chi_lim(long chi_lim_);
